refactor: movido popover de config de curso a su propio componente
